contracts: Efficient encoding of deposit args

+++ packages/contracts/contracts/L1/OptimismPortal.sol
@@ -35,6 +35,6 @@ contract OptimismPortal is DepositFeed, WithdrawalsRelay {
      * depositTransaction() function directly.
      */
     receive() external payable {
-        depositTransaction(msg.sender, msg.value, 30000, false, bytes(""));
+        depositTransaction(msg.sender, uint96(msg.value), 30000, false, bytes(""));
     }
 }
